pyproject.toml
[project]
name = "ragsocket"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk>=4.6.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
